Restart... but complete!
Complete as it relates to the Odin Project's exercise. It's now time to
move forward.

This website I decided change the class scheme to use them more actively
and the result was a much quicker build. I was able to build this site
in roughly 3 hours or so, which is a major improvement over what I was
trying to do before. Additionally, I chose to deviate from the image and
focus less on being pixel-perfect with margins and whatnot so that I
could pump out a full webpage. It's missing images, and I anticipate
that the style will have to be mildly adjusted after adding the images,
but I hope that this code is readable and editable to others!