Add click event listeners attached with alerts
The js file implements a button color change and an alert to display
it's working in listening for the event. Future changes include addind
a drop in form to prompt for the users email address and a thank you
message. Additionally, I'd like to see how to make all buttons change
color upon only pressing one, but for times sake, I need to keep on with
the ODIN Project's curriculum.